- Separate the go parser tests into a separate integration test suite with it's own settings and pytest ini to avoid causing issues with the existing pytest tests
