+++ Taskfile.yml
@@ -127,6 +127,15 @@ tasks:
       - task: backend-up
       - docker compose -f docker-compose.yml exec web sh -c "pytest --basetemp=/home/tdpapp/ {{.PYTEST_ARGS}}"
 
+  backend-pytest-go-integration:
+    desc: 'Run the live Go parser integration suite. E.g: task backend-pytest-go-integration PYTEST_ARGS="-k section1 -vv"'
+    dir: tdrs-backend
+    vars:
+      PYTEST_ARGS: '{{.PYTEST_ARGS | default ""}}'
+    cmds:
+      - task: backend-up
+      - docker compose -f docker-compose.yml exec web sh -c "DJANGO_SETTINGS_MODULE=tdpservice.settings.go_parser_integration DJANGO_CONFIGURATION=GoParserIntegration pytest -c pytest-go-parser.ini --basetemp=/home/tdpapp/ tdpservice/parsers/test/integration {{.PYTEST_ARGS}}"
+
   backend-pytest-cov:
     desc: 'Generate coverage.json file in the backend container. E.g: task backend-pytest-cov PYTEST_ARGS="tdpservice/test/ -s -vv"'
     dir: tdrs-backend
