+++ tdrs-backend/pytest-go-parser.ini
@@ -1,0 +1,6 @@
+[pytest]
+DJANGO_SETTINGS_MODULE = tdpservice.settings.go_parser_integration
+python_files = test_*.py
+testpaths = tdpservice/parsers/test/integration
+markers =
+    go_parser_integration: tests that require the live Go parser service and shared database state
