+++ tdrs-backend/setup.cfg
@@ -60,3 +60,5 @@ DJANGO_SETTINGS_MODULE = tdpservice.settings.local
 python_files = test/test_*.py
 addopts =
     --cov
+norecursedirs =
+    tdpservice/parsers/test/integration
